Look for method definitions marked with a `[RequireNamedArgs]` attribute instead of a special comment.
With the exception of XML Documentation Comments, C# comments don't get emitted into the assembly's metadata.
As a result if a method marked `// [RequireNamedArgs]` from AssemblyA is invoked with positional arguments from AssemblyB
the analyzer doesn't emit any diagnostics for such an invocation. But it should!
Using an attribute seems to be the most reliable approach with regards to the above.