Remove RequireNamedArgs.Vsix
First, there is no plans to distribute the analyzer as a VSIX package.
Second, RequireNamedArgs.Vsix does not work with VS2022 -- "Could not load file or assembly 'FSharp.Core.dll' ...".
After spending two days looking for and trying out various work-arounds (e.g., https://stackoverflow.com/a/29363714/818321) for this issue
I decided that trying to fix the VSIX is just not worth the effort of further investigation.